chore(podman/minio): Restrict mc client permissions
Update permissions for the downloaded MinIO client (mc) binary.
In roles/podman_services/tasks/minio.yml, the mode parameter for
the get_url task fetching mc has been changed from "0755" to "0700".
This change restricts read, write, and execute permissions for the
/usr/local/bin/mc file to only the owner (deploy_user_name), removing
read and execute permissions for the group and others. This enhances
the security posture by limiting access to the executable binary.